onboarding revamp, better rerendering
